Avoid various type casting warnings.

+++ libtiff/tif_dirwrite.c
@@ -294,7 +294,7 @@ TIFFWriteDirectory(TIFF* tif)
 			 */
 			if (dir->tdir_count > 0) {
 				tif->tif_flags |= TIFF_INSUBIFD;
-				tif->tif_nsubifd = dir->tdir_count;
+				tif->tif_nsubifd = (uint16) dir->tdir_count;
 				if (dir->tdir_count > 1)
 					tif->tif_subifdoff = dir->tdir_offset;
 				else
@@ -380,7 +380,7 @@ TIFFWriteNormalTag(TIFF* tif, TIFFDirEntry* dir, const TIFFFieldInfo* fip)
 	u_short wc = (u_short) fip->field_writecount;
 	uint32 wc2;
 
-	dir->tdir_tag = fip->field_tag;
+	dir->tdir_tag = (uint16) fip->field_tag;
 	dir->tdir_type = (u_short) fip->field_type;
 	dir->tdir_count = wc;
 #define	WRITEF(x,y)	x(tif, fip->field_type, fip->field_tag, dir, wc, y)
@@ -525,7 +525,7 @@ TIFFWriteNormalTag(TIFF* tif, TIFFDirEntry* dir, const TIFFFieldInfo* fip)
 static void
 TIFFSetupShortLong(TIFF* tif, ttag_t tag, TIFFDirEntry* dir, uint32 v)
 {
-	dir->tdir_tag = tag;
+	dir->tdir_tag = (uint16) tag;
 	dir->tdir_count = 1;
 	if (v > 0xffffL) {
 		dir->tdir_type = (short) TIFF_LONG;
@@ -625,7 +625,7 @@ TIFFWriteShortTable(TIFF* tif,
 {
 	uint32 i, off;
 
-	dir->tdir_tag = tag;
+	dir->tdir_tag = (uint16) tag;
 	dir->tdir_type = (short) TIFF_SHORT;
 	/* XXX -- yech, fool TIFFWriteData */
 	dir->tdir_count = (uint32) (1L<<tif->tif_dir.td_bitspersample);
@@ -660,7 +660,7 @@ static int
 TIFFWriteShortArray(TIFF* tif,
     TIFFDataType type, ttag_t tag, TIFFDirEntry* dir, uint32 n, uint16* v)
 {
-	dir->tdir_tag = tag;
+	dir->tdir_tag = (uint16) tag;
 	dir->tdir_type = (short) type;
 	dir->tdir_count = n;
 	if (n <= 2) {
@@ -686,7 +686,7 @@ static int
 TIFFWriteLongArray(TIFF* tif,
     TIFFDataType type, ttag_t tag, TIFFDirEntry* dir, uint32 n, uint32* v)
 {
-	dir->tdir_tag = tag;
+	dir->tdir_tag = (uint16) tag;
 	dir->tdir_type = (short) type;
 	dir->tdir_count = n;
 	if (n == 1) {
@@ -708,7 +708,7 @@ TIFFWriteRationalArray(TIFF* tif,
 	uint32* t;
 	int status;
 
-	dir->tdir_tag = tag;
+	dir->tdir_tag = (uint16) tag;
 	dir->tdir_type = (short) type;
 	dir->tdir_count = n;
 	t = (uint32*) _TIFFmalloc(2*n * sizeof (uint32));
@@ -731,7 +731,7 @@ TIFFWriteRationalArray(TIFF* tif,
 			while (fv < 1L<<(31-3) && den < 1L<<(31-3))
 				fv *= 1<<3, den *= 1L<<3;
 		}
-		t[2*i+0] = sign * (fv + 0.5);
+		t[2*i+0] = (uint32) (sign * (fv + 0.5));
 		t[2*i+1] = den;
 	}
 	status = TIFFWriteData(tif, dir, (char *)t);
@@ -743,7 +743,7 @@ static int
 TIFFWriteFloatArray(TIFF* tif,
     TIFFDataType type, ttag_t tag, TIFFDirEntry* dir, uint32 n, float* v)
 {
-	dir->tdir_tag = tag;
+	dir->tdir_tag = (uint16) tag;
 	dir->tdir_type = (short) type;
 	dir->tdir_count = n;
 	TIFFCvtNativeToIEEEFloat(tif, n, v);
@@ -758,7 +758,7 @@ static int
 TIFFWriteDoubleArray(TIFF* tif,
     TIFFDataType type, ttag_t tag, TIFFDirEntry* dir, uint32 n, double* v)
 {
-	dir->tdir_tag = tag;
+	dir->tdir_tag = (uint16) tag;
 	dir->tdir_type = (short) type;
 	dir->tdir_count = n;
 	TIFFCvtNativeToIEEEDouble(tif, n, v);
@@ -787,9 +787,9 @@ TIFFWriteAnyArray(TIFF* tif,
 	switch (type) {
 	case TIFF_BYTE:
 		{ uint8* bp = (uint8*) w;
-		  for (i = 0; i < n; i++)
+		  for (i = 0; i < (int) n; i++)
 			bp[i] = (uint8) v[i];
-		  dir->tdir_tag = tag;
+		  dir->tdir_tag = (uint16) tag;
 		  dir->tdir_type = (short) type;
 		  dir->tdir_count = n;
 		  if (!TIFFWriteByteArray(tif, dir, (char*) bp))
@@ -798,9 +798,9 @@ TIFFWriteAnyArray(TIFF* tif,
 		break;
 	case TIFF_SBYTE:
 		{ int8* bp = (int8*) w;
-		  for (i = 0; i < n; i++)
+		  for (i = 0; i < (int) n; i++)
 			bp[i] = (int8) v[i];
-		  dir->tdir_tag = tag;
+		  dir->tdir_tag = (uint16) tag;
 		  dir->tdir_type = (short) type;
 		  dir->tdir_count = n;
 		  if (!TIFFWriteByteArray(tif, dir, (char*) bp))
@@ -809,7 +809,7 @@ TIFFWriteAnyArray(TIFF* tif,
 		break;
 	case TIFF_SHORT:
 		{ uint16* bp = (uint16*) w;
-		  for (i = 0; i < n; i++)
+		  for (i = 0; i < (int) n; i++)
 			bp[i] = (uint16) v[i];
 		  if (!TIFFWriteShortArray(tif, type, tag, dir, n, (uint16*)bp))
 				goto out;
@@ -817,7 +817,7 @@ TIFFWriteAnyArray(TIFF* tif,
 		break;
 	case TIFF_SSHORT:
 		{ int16* bp = (int16*) w;
-		  for (i = 0; i < n; i++)
+		  for (i = 0; i < (int) n; i++)
 			bp[i] = (int16) v[i];
 		  if (!TIFFWriteShortArray(tif, type, tag, dir, n, (uint16*)bp))
 			goto out;
@@ -825,7 +825,7 @@ TIFFWriteAnyArray(TIFF* tif,
 		break;
 	case TIFF_LONG:
 		{ uint32* bp = (uint32*) w;
-		  for (i = 0; i < n; i++)
+		  for (i = 0; i < (int) n; i++)
 			bp[i] = (uint32) v[i];
 		  if (!TIFFWriteLongArray(tif, type, tag, dir, n, bp))
 			goto out;
@@ -833,7 +833,7 @@ TIFFWriteAnyArray(TIFF* tif,
 		break;
 	case TIFF_SLONG:
 		{ int32* bp = (int32*) w;
-		  for (i = 0; i < n; i++)
+		  for (i = 0; i < (int) n; i++)
 			bp[i] = (int32) v[i];
 		  if (!TIFFWriteLongArray(tif, type, tag, dir, n, (uint32*) bp))
 			goto out;
@@ -841,7 +841,7 @@ TIFFWriteAnyArray(TIFF* tif,
 		break;
 	case TIFF_FLOAT:
 		{ float* bp = (float*) w;
-		  for (i = 0; i < n; i++)
+		  for (i = 0; i < (int) n; i++)
 			bp[i] = (float) v[i];
 		  if (!TIFFWriteFloatArray(tif, type, tag, dir, n, bp))
 			goto out;
